+++ HISTORY.rst
@@ -8,6 +8,7 @@ Version: 1.5.8 - IN DEVELOPMENT
 - #586: removed Python 2.6 support.
 - #591: speedup logging by 28% by using `logging._srcfile = None` trick. This
   avoids calling `calling sys._getframe()` for each log record.
+- #605: added support for Python 3.12.
 
 Version: 1.5.7 - 2022-10-04
 ===========================